+++ packages/codegen/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- Add `viewFns` code generation
+- Expose `<key>Apis` type for Runtime APIs.
+
 ## 0.14.0 - 2025-04-24
 
 ### Changed